Basic Psi elements for expressions. Some work on the inspections.
Changed the Psi for Literals. We have an GoLiteralExpression that can return a literal. That returned literal can be: Int, Bool, String, Char, Float, Imaginary of Identifier. When an identifier literal is parsed inside a const declaration and it's name is iota it will return true when isIota method is called on him.

Moved the UnresolvedSymbol logic into an annotation to be able to test the functionality easy. Changed the return method of the AbstractWholeGoFileAnnoation to return a list instead of an array and updated the clients.

Updated all the psi clients to use the new structure for the basic expressions.
Signed-off-by: Mihai Claudiu Toader <mtoader@gmail.com>