Windows compatibility fixes
GoAppEngineRunningState.java:

It is trying to start dev_appserver.py Python script
through CreateProcess() Windows API call, which will not work. We must set
ExePath to python.exe, and set the script file name as the 1st argument.

GoSdkUtil.java:

RE_OS_MATCHER was defined incorrectly and never worked - the
program thus thought that it is always on Linux. Question marks were needed at
the start and end of RegEx expression.

Around line 97, goCommand was set to path + "/bin/go" and then a file exist
check was done. Well, on Windows, you must check for   + "/bin/go.exe" instead...

At line 249 we were getting an exception because of using VfsUtil.findFileByURL(),
and the whole URL construction and conversion back to path is screwed up for
file objects on Windows - for a path like D:/somedir/someother, the URL makes
D the "host" - which is wrong, and sets path to "/somedir/someother", then good
luck finding it on the file system! Fixed by using VfsUtil.findFileByIoFile()
instead.