Fix regex for cases when parsing minimal GDB response
It is possible to receive an extremely minimal response from GDB, such as:

`{addr="0x00002033",func="main.main",args=[]}`

Previously, if an args parameter was detected, the regex required that it be
present and followed by other parameters.  In the above snippet, the regex
failed to match and resulted in an eventual OutOfBoundsArrayException.

This fix makes the trailing comma in the args regex optional