Fix GoInspectionTestCase.doTestWithDirectory() to work on Windows.
Because FileUtil.getRelativePath(File, File) uses the character specified
by File.pathSeparator for the returned file path, the returned paths
use '/' on Linux and '\\' on Windows.

This causes a number of tests to fail on Windows, because
myFixture.configureFromTempProjectFile(fileName)
requires '/' as the path separator.

The issue can be fixed by normalizing the path before passing it to
configureFromTempProjectFile (as it is done in this commit).