consider aliases and path names when looking for dupe imports
Previously, the duplicate import detector would ignore import aliases or last
path components for the purposes of detecting duplicates.

e.g.

import "a/pack"
import "b/pack"

would not be an error. Similarly

import a "pack1"
import a "pack2"

would also not be an error

This change also adds the GoUnusedImportDeclaration inspector to the list of
inspectors for GoHighlightTest. A few of the test cases had to be changed to
avoid errors raised by this inspector.