add an inspection for StructTag
This is an incomplete implementation of
https://github.com/go-lang-plugin-org/go-lang-idea-plugin/issues/1983. It will
raise warnings if it detects a tag value that does not conform to the Go
StructTag convention described by https://golang.org/pkg/reflect/#StructTag.

Currently, the inspector does not validate that the tag value strings are valid
and is therefore overly permissive.

Its implementation is essentially ported from validateStructTag in
https://github.com/golang/tools/blob/master/cmd/vet/structtag.go.