fix 'continue not in a loop' inspector
Previously the inspector would look for the presence of a GoForStatement
parent of a continue statement. This did not take into account the
presence of 'go' or 'defer' calls, e.g.

for {
    defer func() {
        continue // no error was raised, should be an error
    }()
}

The discussion at
https://github.com/go-lang-plugin-org/go-lang-idea-plugin/pull/2024
reminded me that the continue inspection suffers from this problem.
(cherry picked from commit a58f2b0)